feat(FOROME-1032): create button css module, with all styles
