+++ src/components/export-report-button.tsx
@@ -3,9 +3,10 @@ import cn from 'classnames'
 import { observer } from 'mobx-react-lite'
 
 import { t } from '@i18n'
+import { theme } from '@theme'
 import operationsStore from '@store/operations'
 import mainTableStore from '@store/ws/main-table.store'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Icon } from '@ui/icon'
 import { MainTableDataCy } from './data-testid/main-table.cy'
 import { Loader } from './loader'
@@ -19,11 +20,11 @@ interface Props {
 export const ExportReportButton = observer(
   ({ isOpen, refEl, ...rest }: Props): ReactElement => {
     const { variantCounts } = mainTableStore.fixedStatAmount
-    const areVariantsZero = variantCounts === 0
+    const disabled = !variantCounts
 
     return (
       <Button
-        disabled={areVariantsZero}
+        disabled={disabled}
         text={
           operationsStore.isExportingReport ? (
             <Loader size="xs" color="white" />
@@ -34,13 +35,13 @@ export const ExportReportButton = observer(
         dataTestId={MainTableDataCy.exportReport}
         refEl={refEl}
         size="sm"
+        variant="primary-dark"
         prepend={<Icon name="Export" />}
         onClick={rest.onClick}
         style={{
-          width: '157px',
           pointerEvents: operationsStore.isExportingReport ? 'none' : 'inherit',
+          backgroundColor: !disabled && theme('colors.blue.secondary'),
         }}
-        className="bg-blue-secondary hover:bg-blue-lighter active:bg-blue-lighter"
         textSize="sm"
         append={
           <Icon
