+++ src/components/undo-redo-buttons.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react'
 
-import { Button } from '@ui/button'
+import { Button } from '../ui/button/button'
 import { Icon } from '@ui/icon'
 
 export const UndoRedoButtons = (): ReactElement => (
