+++ src/components/variant/ui/drawer-note/drawer-note.tsx
@@ -9,7 +9,7 @@ import { useOutsideClick } from '@core/hooks/use-outside-click'
 import { useResizeTextAreaHeight } from '@core/hooks/use-resize-text-area-height'
 import { t } from '@i18n'
 import variantStore from '@store/ws/variant'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Icon } from '@ui/icon'
 import { VariantDrawerDataCy } from '@components/data-testid/variant-drawer.cy'
 import { PopperButton } from '@components/popper-button'
