+++ src/pages/error/error.tsx
@@ -2,7 +2,7 @@ import { useHistory } from 'react-router-dom'
 
 import { t } from '@i18n'
 import { Routes } from '@router/routes.enum'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 
 export const ErrorPage = () => {
   const history = useHistory()
