+++ src/pages/filter/dtree/components/filter-control-dtree.tsx
@@ -8,7 +8,7 @@ import { ActionFilterEnum } from '@core/enum/action-filter.enum'
 import { t } from '@i18n'
 import datasetStore from '@store/dataset/dataset'
 import dtreeStore from '@store/dtree'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { DropDown } from '@ui/dropdown'
 import { Input } from '@ui/input'
 import { DecisionTreesMenuDataCy } from '@components/data-testid/decision-tree-menu.cy'
