+++ src/pages/filter/dtree/components/modals/components/modal-inheritance-mode/components/inheritance-mode-content.tsx
@@ -5,7 +5,7 @@ import { ModeTypes } from '@core/enum/mode-types-enum'
 import { SubKinds } from '@core/enum/sub-kinds-enum'
 import { t } from '@i18n'
 import dtreeStore from '@store/dtree'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Checkbox } from '@ui/checkbox/checkbox'
 import { AllNotMods } from '@pages/filter/dtree/components/query-builder/ui/all-not-mods'
 import { DisabledVariantsAmount } from '@pages/filter/dtree/components/query-builder/ui/disabled-variants-amount'
