+++ src/pages/filter/dtree/components/modals/components/modal-save-dataset.tsx
@@ -15,7 +15,7 @@ import operations from '@store/operations'
 import mainTableStore from '@store/ws/main-table.store'
 import zoneStore from '@store/ws/zone'
 import { Routes } from '@router/routes.enum'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Attention } from '@ui/icons/attention'
 import { Input } from '@ui/input'
 import { DecisionTreesMenuDataCy } from '@components/data-testid/decision-tree-menu.cy'
