+++ src/pages/filter/dtree/components/modals/components/modal-text-editor.tsx
@@ -9,7 +9,7 @@ import { LocalStoreManager } from '@core/storage-management/local-store-manager'
 import { t } from '@i18n'
 import datasetStore from '@store/dataset/dataset'
 import dtreeStore from '@store/dtree'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { DecisionTreeModalDataCy } from '@components/data-testid/decision-tree-modal.cy'
 import Editor from '@monaco-editor/react'
 import { IDtreeCheck } from '@service-providers/decision-trees/decision-trees.interface'
