+++ src/pages/filter/dtree/components/modals/components/ui/select-modal-buttons.tsx
@@ -5,7 +5,7 @@ import { observer } from 'mobx-react-lite'
 import { ActionType } from '@declarations'
 import { ModalSources } from '@core/enum/modal-sources'
 import { t } from '@i18n'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Icon } from '@ui/icon'
 import { DecisionTreeModalDataCy } from '@components/data-testid/decision-tree-modal.cy'
 import modalsVisibilityStore from '../../modals-visibility-store'
