+++ src/pages/filter/refiner/components/middle-column/panels/enum-panel.tsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite'
 import { ModeTypes } from '@core/enum/mode-types-enum'
 import { t } from '@i18n'
 import filterStore from '@store/filter'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Pagintaion } from '@components/pagintaion'
 import { QueryBuilderSearch } from '@pages/filter/dtree/components/query-builder/query-builder-search'
 import { AllNotMods } from '@pages/filter/dtree/components/query-builder/ui/all-not-mods'
