+++ src/pages/filter/refiner/components/middle-column/panels/function-panel/components/compound-request/control-buttons.tsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite'
 
 import { RequestBlockOperations } from '@core/enum/request-block-operations'
 import filterStore from '@store/filter'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import compoundRequestStore from './compound-request.store'
 
 interface IControlButtonsProps {
