+++ src/pages/filter/refiner/components/middle-column/panels/function-panel/components/panelButtons.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react-lite'
 
 import { t } from '@i18n'
 import filterStore from '@store/filter'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 
 interface IPanelButtons {
   onSubmit: () => void
