+++ src/pages/filter/refiner/components/right-column/query-selected.tsx
@@ -9,7 +9,7 @@ import datasetStore from '@store/dataset/dataset'
 import dtreeStore from '@store/dtree'
 import filterStore from '@store/filter'
 import { Routes } from '@router/routes.enum'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Loader } from '@components/loader'
 import { FilterRefinerStatCounts } from '@pages/filter/refiner/components/right-column/filter-refiner-stat-counts'
 import { showToast } from '@utils/notifications/showToast'
