+++ src/pages/main/components/handle-dataset/handle-dataset-button.tsx
@@ -1,6 +1,6 @@
 import { MouseEvent } from 'react'
 
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Icon } from '@ui/icon'
 
 interface IHandleDatasetButtonProps {
@@ -19,7 +19,8 @@ export const HandleDatasetButton = ({
       refEl={refEl}
       onClick={onClick}
       className="rounded"
-      prepend={<Icon name="Ellipsis" />}
+      size="md"
+      icon={<Icon name="Ellipsis" />}
       onMouseUp={onMouseUp}
       style={{
         width: '36px',
