+++ src/pages/not-found/not-found.page.tsx
@@ -6,7 +6,7 @@ import styled from 'styled-components'
 import { t } from '@i18n'
 import { theme } from '@theme'
 import { Routes } from '@router/routes.enum'
-import { Button } from '@ui/button'
+import { Button } from '@ui/button/button'
 import { Icon } from '@ui/icon'
 import { Header } from '@components/header'
 import { ErrorPage } from '@pages/error/error'
